add months delay; improve performance of loop; change type from decimal to double

+++ LoanPortfolioCore/FlatFileModels/LoanOutput.cs
@@ -9,10 +9,10 @@ namespace LoanPortfolioCore.FlatFileModels
     class LoanOutput
     {
             public int LoanId { get; set; }
-            public decimal Principal { get; set; }
+            public double Principal { get; set; }
             public double Rate { get; set; } //annual
             public int TermInMonths { get; set; }
             public string LoanName { get; set; }
-            public decimal MinPayment { get; set; }
+            public double MinPayment { get; set; }
     }
 }
